# perf(ilp): optimize ingestion for tables with large number of columns