--- a/core/src/main/java/io/questdb/std/str/Utf8Sequence.java
+++ b/core/src/main/java/io/questdb/std/str/Utf8Sequence.java
@@ -28,7 +28,8 @@ import io.questdb.cairo.VarcharTypeDriver;
 import io.questdb.std.Unsafe;
 import org.jetbrains.annotations.NotNull;
 
-import static io.questdb.cairo.VarcharTypeDriver.*;
+import static io.questdb.cairo.VarcharTypeDriver.VARCHAR_INLINED_PREFIX_MASK;
+import static io.questdb.cairo.VarcharTypeDriver.VARCHAR_MAX_BYTES_FULLY_INLINED;
 
 /**
  * A sequence of UTF-8 bytes.
@@ -54,6 +55,36 @@ public interface Utf8Sequence {
     byte byteAt(int index);
 
     /**
+     * Called as a part of equality check that has already ensured the two strings
+     * have the same byte size. This is especially relevant when comparing two values
+     * from a VARCHAR column: same size guarantees they are either both inlined or
+     * both not inlined, which means the same `Utf8Sequence` implementation is on
+     * both sides.
+     */
+    default boolean equalsAssumingSameSize(Utf8Sequence other) {
+        int i = 0;
+        for (int n = size() - 7; i < n; i += 8) {
+            if (longAt(i) != other.longAt(i)) {
+                return false;
+            }
+        }
+        for (int n = size(); i < n; i++) {
+            if (byteAt(i) != other.byteAt(i)) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    /**
+     * Returns `true` if it's guaranteed that the contents of this UTF-8 sequence are
+     * all ASCII characters. Returning `false` does not guarantee anything.
+     */
+    default boolean isAscii() {
+        return false;
+    }
+
+    /**
      * Returns eight bytes of the UTF-8 sequence located at the provided
      * byte offset, packed into a single `long` value. The bytes are arranged
      * in little-endian order. The method does not check bounds and will
@@ -71,47 +102,6 @@ public interface Utf8Sequence {
     }
 
     /**
-     * Returns the first 6 bytes of this UTF-8 sequence packed into a zero-padded long
-     * value, in little-endian order. This prefix is stored inline in the auxiliary vector
-     * of a VARCHAR column, so asking for it is a matter of optimized data access. This is
-     * not a general access method, it shouldn't be called except when looking to optimize
-     * the access of the VARCHAR column.
-     *
-     * This method should be called only on a UTF-8 sequence longer than {@value
-     * VarcharTypeDriver#VARCHAR_MAX_BYTES_FULLY_INLINED}. VARCHAR values shorter than that
-     * are stored in a different format.
-     */
-    default long zeroPaddedSixPrefix() {
-        assert size() > VARCHAR_MAX_BYTES_FULLY_INLINED
-                : String.format("size %,d <= %d", size(), VARCHAR_MAX_BYTES_FULLY_INLINED);
-        return longAt(0) & VARCHAR_INLINED_PREFIX_MASK;
-    }
-
-    /**
-     * Called as a part of equality check that has already ensured the two strings
-     * have the same byte size. This is especially relevant when comparing two values
-     * from a VARCHAR column: same size guarantees they are either both inlined or
-     * both not inlined, which means the same `Utf8Sequence` implementation is on
-     * both sides.
-     */
-    default boolean equalsAssumingSameSize(Utf8Sequence other) {
-        for (int i = 0, n = size(); i < n; i++) {
-            if (byteAt(i) != other.byteAt(i)) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    /**
-     * Returns `true` if it's guaranteed that the contents of this UTF-8 sequence are
-     * all ASCII characters. Returning `false` does not guarantee anything.
-     */
-    default boolean isAscii() {
-        return false;
-    }
-
-    /**
      * Number of bytes in the string.
      * <p>
      * This is NOT the number of 16-bit chars or code points in the string.
@@ -120,8 +110,29 @@ public interface Utf8Sequence {
     int size();
 
     default void writeTo(long addr, int lo, int hi) {
-        for (int i = lo; i < hi; i++) {
-            Unsafe.getUnsafe().putByte(addr++, byteAt(i));
+        int i = lo;
+        for (int n = hi - 7; i < n; i += 8, addr += 8) {
+            Unsafe.getUnsafe().putLong(addr, longAt(i));
         }
+        for (; i < hi; i++, addr++) {
+            Unsafe.getUnsafe().putByte(addr, byteAt(i));
+        }
+    }
+
+    /**
+     * Returns the first 6 bytes of this UTF-8 sequence packed into a zero-padded long
+     * value, in little-endian order. This prefix is stored inline in the auxiliary vector
+     * of a VARCHAR column, so asking for it is a matter of optimized data access. This is
+     * not a general access method, it shouldn't be called except when looking to optimize
+     * the access of the VARCHAR column.
+     * <p>
+     * This method should be called only on a UTF-8 sequence longer than {@value
+     * VarcharTypeDriver#VARCHAR_MAX_BYTES_FULLY_INLINED}. VARCHAR values shorter than that
+     * are stored in a different format.
+     */
+    default long zeroPaddedSixPrefix() {
+        assert size() > VARCHAR_MAX_BYTES_FULLY_INLINED
+                : String.format("size %,d <= %d", size(), VARCHAR_MAX_BYTES_FULLY_INLINED);
+        return longAt(0) & VARCHAR_INLINED_PREFIX_MASK;
     }
 }