--- a/core/src/main/java/io/questdb/std/str/Utf8StringSink.java
+++ b/core/src/main/java/io/questdb/std/str/Utf8StringSink.java
@@ -82,6 +82,11 @@ public class Utf8StringSink implements MutableUtf8Sink {
     }
 
     @Override
+    public long longAt(int offset) {
+        return Unsafe.byteArrayGetLong(buffer, offset);
+    }
+
+    @Override
     public Utf8StringSink put(@Nullable Utf8Sequence us) {
         if (us != null) {
             ascii &= us.isAscii();