Includes the following:
* Optimizes `TableUpdateDetails#clearProcessedColumns()` method
* Makes use of `longAt()` method in the default `equalsAssumingSameSize()` and `writeTo()` method implementations in `Utf8Sequence`

### Benchmarks

Loader which ingests a table with 1k long columns: https://gist.github.com/puzpuzpuz/fa551d055649af685328792e45fe73d9

Before:
```bash
$ go run main.go --workers=1
[0] Sender started
Throughput: 2800 rows/s
Throughput: 2900 rows/s
Throughput: 2800 rows/s
```

After:
```bash
$ go run main.go --workers=1
[0] Sender started
Throughput: 3900 rows/s
Throughput: 4000 rows/s
Throughput: 3900 rows/s
```